Reuse shared parseSlashCommand and hoist findAwaitingPermissionThreadId
Lark and Weixin gateways each carried private copies of parseSlashCommand
that already existed as an exported helper in local-core-slash-commands,
plus near-identical findAwaitingPermissionThreadId iterating the same
threadRouting map. Drop the copies and hoist one protected method to
BaseChannelGateway with an optional platformKey filter so both subclass
behaviors (Lark: no filter; Weixin: hard-coded 'weixin' default) are
preserved exactly.
